Create List - House Party
Write a program that keeps track of the guests that are going to a house party. 
On the first input line, you are going to receive how many commands you are going to have. On the next lines you are going to receive some of the following inputs:
•	"{name} is going!"
•	"{name} is not going!"
If you receive the first type of input, you have to add the person if he/she is not on the list. (If he/she is in the list print on the console: "{name} is already in the list!"). If you receive the second type of input, you have to remove the person if he/she is on the list. (If not print: "{name} is not in the list!"). At the end print all guests.

Examples:
Input:
4
Allie is going!
George is going!
John is not going!
George is not going!

Output:
John is not in the list!
Allie